refactor: split overlay sections, choose-tech rows, and shell chrome (#3878)
Decompose province overlay section helpers, technology choose-tech dialog
rows, intervention scrim shell, train civilians unit row, and main-menu body
into library parts; update SPEC and architecture pin tests for new paths.

Refs #3878
